Refactor: 기존에 테스트용으로 만든 Item 도메인 변경 Refactor
