+++ src/main/java/programmers/coffee/domain/item/domain/Item.java
@@ -8,5 +8,5 @@ public class Item {
     private String itemName;
     private String description;
     private int price;
-    private boolean soldOut;
+    private int stockCnt;
 }
