+++ src/main/java/programmers/coffee/domain/order/dto/OrderItemDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Setter;
-import programmers.coffee.domain.order.domain.Item;
+import programmers.coffee.domain.item.domain.Item; //
 import programmers.coffee.domain.order.domain.OrderItem;
 
 @Getter
@@ -32,8 +32,4 @@ public class OrderItemDto {
         orderItem.setItemName(item.getItemName());
         return orderItem;
     }
-
-
 }
-
-
