+++ src/main/resources/mapper/item-mapper.xml
@@ -1,33 +1,30 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="programmers.coffee.domain.item.repository.ItemRepository">
 
-  <resultMap id="itemResultMap" type="programmers.coffee.domain.order.domain.Item">
+  <resultMap id="itemResultMap" type="programmers.coffee.domain.item.domain.Item">
     <id property="itemId" column="item_id"/>
     <result property="itemName" column="item_name"/>
+    <result property="description" column="description"/>
     <result property="price" column="price"/>
+    <result property="soldOut" column="sold_out"/>
     <result property="stockCnt" column="stock_cnt"/>
   </resultMap>
 
-  <select id="findById" resultMap="itemResultMap">
-    SELECT * FROM item
-    WHERE item_id = #{itemId}
-  <select id="findAll" resultType="programmers.coffee.domain.item.domain.Item">
-    SELECT * FROM item;
+  <select id="findById" parameterType="long" resultMap="itemResultMap">
+    SELECT * FROM item WHERE item_id = #{itemId}
   </select>
 
   <select id="findAll" resultMap="itemResultMap">
     SELECT * FROM item
-  <select id="findById" parameterType="long" resultType="programmers.coffee.domain.item.domain.Item">
-    SELECT * FROM item WHERE id = #{id};
   </select>
 
   <insert id="insertItem" parameterType="programmers.coffee.domain.item.domain.Item">
-    INSERT INTO item(item_name, description, price, sold_out)
-    VALUES(#{itemName}, #{description}, #{price}, #{soldOut});
+    INSERT INTO item(item_name, description, price, sold_out, stock_cnt)
+    VALUES(#{itemName}, #{description}, #{price}, #{soldOut}, #{stockCnt})
   </insert>
 
 </mapper>
